+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-relay-offline",
-  "version": "0.5.0",
+  "version": "0.6.0-a3",
   "keywords": [
     "graphql",
     "relay",
@@ -24,7 +24,7 @@
   "dependencies": {
     "@babel/runtime": "^7.0.0",
     "@wora/detect-network": "0.1.0",
-    "@wora/cache-persist": "0.4.1",
+    "@wora/cache-persist": "0.4.3",
     "fbjs": "^1.0.0",
     "nullthrows": "^1.1.0",
     "uuid": "3.3.2"
