+++ src/index.ts
@@ -1,6 +1,7 @@
 export {
   ReactRelayContext,
   applyOptimisticMutation,
+  commitMutation,
   commitLocalUpdate,
   createFragmentContainer,
   createPaginationContainer,
@@ -9,7 +10,6 @@ export {
   graphql,
   requestSubscription,
 } from 'react-relay';
-
 export {
   $FragmentRef,
   RelayFragmentContainer,
@@ -40,8 +40,6 @@ export {default as Environment} from './runtime/RelayModernEnvironment';
 export {default as Store} from './runtime/Store';
 export {default as RecordSource} from './runtime/RecordSource';
 export {default as OfflineStore} from './runtime/StoreOffline';
-export {default as useMutation} from './hooks/useMutation';
-export {default as commitMutation} from './commitMutation';
 export { NetInfo } from '@wora/detect-network';
 export { useNetInfo } from '@wora/detect-network';
 export { useIsConnected } from '@wora/detect-network';
