+++ src/runtime/StoreOffline.ts
@@ -2,6 +2,24 @@ import Cache, { CacheOptions } from "@wora/cache-persist";
 import { v4 as uuid } from "uuid";
 import { Disposable, Network, } from 'relay-runtime/lib/RelayStoreTypes';
 import RelayModernEnvironment from './RelayModernEnvironment';
+import { ROOT_TYPE } from 'relay-runtime/lib/RelayStoreUtils';
+import * as RelayInMemoryRecordSource from 'relay-runtime/lib/RelayInMemoryRecordSource';
+import * as RelayModernRecord from 'relay-runtime/lib/RelayModernRecord';
+import * as RelayResponseNormalizer from 'relay-runtime/lib/RelayResponseNormalizer';
+import { GetDataID } from 'relay-runtime/lib/RelayResponseNormalizer';
+import * as ErrorUtils from 'fbjs/lib/ErrorUtils';
+import * as RelayRecordSourceMutator from 'relay-runtime/lib/RelayRecordSourceMutator';
+import * as RelayRecordSourceProxy from 'relay-runtime/lib/RelayRecordSourceProxy';
+import * as RelayReader from 'relay-runtime/lib/RelayReader';
+import * as normalizeRelayPayload from 'relay-runtime/lib/normalizeRelayPayload';
+import { GraphQLResponseWithData } from 'relay-runtime/lib/RelayNetworkTypes';
+import {
+    NormalizationSelector,
+    RelayResponsePayload
+} from 'relay-runtime/lib/RelayStoreTypes';
+
+import { Observable as RelayObservable } from 'relay-runtime';
+
 
 type Subscription = {
     callback: (state) => void,
@@ -10,8 +28,8 @@ type Subscription = {
 export type OfflineOptions = {
     manualExecution?: boolean;
     network?: Network;
-    onComplete?: (options: {id: string, offlinePayload: any, snapshot: any}) => boolean;
-    onDiscard?: (options: {id: string, offlinePayload: any, error: any}) => boolean;
+    onComplete?: (options: { id: string, offlinePayload: any, snapshot: any }) => boolean;
+    onDiscard?: (options: { id: string, offlinePayload: any, error: any }) => boolean;
     //onDispatch?: (request: any) => any;
 }
 
@@ -50,20 +68,18 @@ class RelayStoreOffline {
 
     }
 
+    public purge(): Promise<boolean> {
+        return this._cache.purge().then(purged => {
+            this.notify();
+            return purged;
+        });
+    }
+
     public restore(): Promise<Cache> {
         return this._cache.restore().then(cache => {
             this.notify();
             return cache;
         });
-    }
-
-    public dispatch(request) {
-        //const { onDispatch } = this._offlineOptions;
-        const fetchTime = Date.now();
-        const id = uuid();
-        const promiseSet = this._cache.set(id, { request, fetchTime });//, extra: onDispatch(request) });
-        this.notify();
-        return promiseSet;
     }
 
     public subscribe(
@@ -124,12 +140,12 @@ class RelayStoreOffline {
         ).subscribe({
             complete: () => {
                 const snapshot = this._environment.lookup(operation.fragment, operation);
-                if (onComplete( {id, offlinePayload: payload, snapshot: (snapshot.data as any) }) ) {
+                if (onComplete({ id, offlinePayload: payload, snapshot: (snapshot.data as any) })) {
                     this._cache.remove(id);
                 }
             },
             error: (error, isUncaughtThrownError?: boolean) => {
-                if (onDiscard({id, offlinePayload: payload, error }))
+                if (onDiscard({ id, offlinePayload: payload, error }))
                     this.discardMutation(id, payload);
             },
             /*next: response => {
@@ -156,7 +172,147 @@ class RelayStoreOffline {
         return source;
     }
 
+    publish(environment, mutationOptions) {
+        return RelayObservable.create(sink => {
+            const {
+                operation,
+                optimisticResponse,
+                optimisticUpdater,
+                updater,
+                uploadables,
+            } = mutationOptions;
+
+
+            const backup = new RelayInMemoryRecordSource();
+            let sinkPublish = new RelayInMemoryRecordSource()
+            if (optimisticResponse || optimisticUpdater) {
+                const sink = new RelayInMemoryRecordSource();
+                const mutator = new RelayRecordSourceMutator(
+                    environment.getStore().getSource(),
+                    sink,
+                    backup
+                );
+                const store = new RelayRecordSourceProxy(mutator, environment._getDataID);
+                const response = optimisticResponse || null;
+                const selectorStoreUpdater = optimisticUpdater;
+                const selectorStore = store.commitPayload(operation, response);
+                // TODO: Fix commitPayload so we don't have to run normalize twice
+                let selectorData, source;
+                if (response) {
+                    ({ source } = normalizeRelayPayload(
+                        operation.root,
+                        response,
+                        null,
+                        { getDataID: environment._getDataID },
+                    ));
+                    selectorData = RelayReader.read(source, operation.fragment, operation).data;
+                }
+                selectorStoreUpdater &&
+                    ErrorUtils.applyWithGuard(
+                        selectorStoreUpdater,
+                        null,
+                        [selectorStore, selectorData],
+                        null,
+                        'RelayPublishQueue:applyUpdates',
+                    );
+            }
+            if (optimisticResponse) {
+                var payload = normalizeResponse({ data: optimisticResponse },
+                    operation.root,
+                    ROOT_TYPE,
+                    [],
+                    environment._getDataID,
+                );
+                const { fieldPayloads, source } = payload;
+                // updater only for configs
+                sinkPublish = environment._publishQueue._getSourceFromPayload({ fieldPayloads, operation, source, updater });
+                //environment._publishQueue.commitPayload(operation, payload, configs ? updater : null);
+                //environment._publishQueue.run();
+
+
+            }
+            /*
+            //environment.
+            //environment.applyUpdate(optimisticUpdate);*/
+
+            if (!optimisticResponse && optimisticUpdater) {
+                const mutator = new RelayRecordSourceMutator(
+                    environment.getStore().getSource(),
+                    sinkPublish,
+                );
+                const store = new RelayRecordSourceProxy(mutator, environment._getDataID);
+                ErrorUtils.applyWithGuard(
+                    optimisticUpdater,
+                    null,
+                    [store],
+                    null,
+                    'RelayPublishQueue:commitUpdaters',
+                );
+                /*    
+                if (optimisticUpdater != null) {
+                    optimisticUpdater(environment.getStore());
+                }
+                if (updater) {
+                    updater(environment.getStore())
+                }*/
+            }
+
+            
+
+
+            //const { onDispatch } = this._offlineOptions;
+            const fetchTime = Date.now();
+            const id = uuid();
+            const request = {
+                operation,
+                optimisticResponse,
+                uploadables,
+                backup,
+                sinkPublish
+            };
+            this.notify();
+            this._cache.set(id, { request, fetchTime }).then(() => {
+                environment.getStore().publish(sinkPublish);
+                environment.getStore().notify();
+                sink.next({
+                    id,
+                    request,
+                    fetchTime
+                });
+                sink.complete();
+            }).catch(error => {
+                sink.error(error, true)
+            });
+        });
+    }
+
 }
 
+
+function normalizeResponse(
+    response: GraphQLResponseWithData,
+    selector: NormalizationSelector,
+    typeName: string,
+    path: ReadonlyArray<string>,
+    getDataID: GetDataID,
+): RelayResponsePayload {
+    const { data, errors } = response;
+    const source = new RelayInMemoryRecordSource();
+    const record = RelayModernRecord.create(selector.dataID, typeName);
+    source.set(selector.dataID, record);
+    const normalizeResult = RelayResponseNormalizer.normalize(
+        source,
+        selector,
+        data,
+        { handleStrippedNulls: true, path, getDataID },
+    );
+    return {
+        errors,
+        incrementalPlaceholders: normalizeResult.incrementalPlaceholders,
+        fieldPayloads: normalizeResult.fieldPayloads,
+        moduleImportPayloads: normalizeResult.moduleImportPayloads,
+        source,
+    };
+}
 
 export default RelayStoreOffline;
